kernel: Boost DDR bus when zygote-forked process becomes top app
Boost to the max for 1 second whenever the top app changes, which
improves app launch speeds and addresses jitter when switching between
apps. A check to make sure that the top-app's parent is zygote ensures
that a user-facing app is indeed what's added to the top app task group,
since app processes are forked from zygote.

Change-Id: Ia3650af54d356de335832c4c5e2576705fe09da1
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: kyvangka1610 <kyvangka2002@gmail.com>