devfreq_boost: Track input event times and expose timeout helper
Sometimes we may want to limit certain boost kicks to only take effect
within a certain timeout after receiving the last input event to save
power. For example, we may want to boost when an intensive process
forks to improve fluidity, but boosting when the fork was not initiated
by input is wasteful.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: dracarys18 <karthihegde010@gmail.com>
Signed-off-by: kyvangka1610 <kyvangka2002@gmail.com>